Split manager up into manager and transitioner
